taichu-demo/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
altair=5.3.0=pypi_0
attrs=23.2.0=pypi_0
blinker=1.8.2=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
certifi=2024.6.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
file-chat-input=0.0.2=pypi_0
filetype=1.2.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
idna=3.7=pypi_0
jinja2=3.1.4=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
mdurl=0.1.2=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.26.4=pypi_0
openssl=3.0.13=h7f8727e_2
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pillow=10.3.0=pypi_0
pip=24.0=py311h06a4308_0
protobuf=4.25.3=pypi_0
pyarrow=16.1.0=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.18.0=pypi_0
python=3.11.9=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.1=pypi_0
setuptools=69.5.1=py311h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.45.3=h5eee18b_0
streamlit=1.35.0=pypi_0
streamlit-float=0.3.4=pypi_0
tenacity=8.3.0=pypi_0
tk=8.6.14=h39e8969_0
toml=0.10.2=pypi_0
toolz=0.12.1=pypi_0
tornado=6.4=pypi_0
typing-extensions=4.12.1=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
watchdog=4.0.1=pypi_0
wheel=0.43.0=py311h06a4308_0
xz=5.4.6=h5eee18b_1
zlib=1.2.13=h5eee18b_1
